refactor(grovedb): dedupe shared aggregate-proof helpers into aggregate_common
The three aggregate-axis verifier modules each carried byte-identical
copies of four helpers (modulo a single axis-label substring in the
diagnostic strings). This was 4 x 3 = 12 copies of essentially the
same logic, and `expect_merk_bytes` was even called out by a reviewer
as obvious duplication.

Centralized into a new `grovedb/src/operations/proof/aggregate_common.rs`:
- `OuterMatch` — pure type, identical across axes.
- `verify_single_key_layer_proof_v0` — identical across axes.
- `expect_merk_bytes` — takes an `axis_label: &'static str` so the
  per-axis prefix in the rejection message is preserved.
- `execute_carrier_layer_proof` — same axis-label parameterization.

Each axis's `helpers.rs` now:
- Re-exports `OuterMatch` + `verify_single_key_layer_proof_v0` so
  existing callers in `leaf_chain.rs` / `per_key.rs` keep their
  `use super::helpers::*` imports working with zero changes.
- Defines a one-line `const AXIS_LABEL: &str` and thin wrappers around
  the shared `expect_merk_bytes` / `execute_carrier_layer_proof` that
  pass the label. Caller signatures and error strings are byte-for-byte
  preserved.
- Keeps the genuinely axis-specific helpers (`verify_*_leaf` with
  different return types per axis, `enforce_lower_chain` with
  different terminal-type acceptance sets).

Net diff: -158 lines, but the real win is structural — future
changes to the shared logic propagate automatically rather than
requiring three parallel updates.

Behavior preserved: all 272 existing aggregate/query tests still pass.
